Rename 'components/DataViewDataSetPanel.vue' to 'components/index/_shared/DataView/DataSetPanel.vue'
